// lib/api/generated/model/wingProfile.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { WingProfileCity } from './wingProfileCity';
import type { WingProfileDatingStatus } from './wingProfileDatingStatus';
import type { WingProfileGender } from './wingProfileGender';
import type { WingProfileInterestsItem } from './wingProfileInterestsItem';

export interface WingProfile {
  profileId: string;
  userId: string;
  chosenName: string;
  /** @nullable */
  gender: WingProfileGender;
  age: number;
  city: WingProfileCity;
  /** @nullable */
  bio: string | null;
  datingStatus: WingProfileDatingStatus;
  interests: WingProfileInterestsItem[];
  /** @nullable */
  firstPhoto: string | null;
}